Added dynamic ponderomotive steps

+++ src/laser_electron.c
@@ -57,7 +57,10 @@ void simulate(struct parameters *param, void (*compute_function)(double *restric
 					higuera_cary_step(&p[k].u[0], dt, l);
 				
 				if(output_mode == 0 && i % substeps == 0) {
-					int idx = id * U_SIZE * steps * num / (substeps * core_num) + (k - initial_idx) * U_SIZE * steps / substeps + i * U_SIZE / substeps;
+					unsigned long idx = (unsigned long)id * U_SIZE * steps * num / (substeps * core_num)
+						+ (unsigned long)(k - initial_idx) * U_SIZE * steps / substeps
+						+ (unsigned long)i * U_SIZE / substeps;
+					
 					memcpy(&out_chunk[idx], &p[k].u[0], U_SIZE * sizeof(double));
 					#pragma omp master
 					{
